Got the buttons to show right pages. It's not hiding the old pages though
